Adjust logging infra in shared crate
