+++ proxy_agent_shared/src/lib.rs
@@ -2,11 +2,10 @@
 // SPDX-License-Identifier: MIT
 
 pub mod error;
-pub mod logger_manager;
+pub mod logger;
 pub mod misc_helpers;
 pub mod proxy_agent_aggregate_status;
 pub mod result;
-pub mod rolling_logger;
 pub mod service;
 pub mod telemetry;
 pub mod version;
